+++ bin/check-rabbitmq-consumers.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 #  encoding: UTF-8
-#
+
 # Check RabbitMQ consumers
 # ===
 #
@@ -65,6 +65,12 @@ class CheckRabbitMQConsumers < Sensu::Plugin::Check::CLI
          long: '--exclude queue_name',
          proc: proc { |q| q.split(',') }
 
+  option :regex,
+         description: 'Treat the --queue flag as a regular expression.',
+         long: '--regex',
+         boolean: true,
+         default: false
+
   option :warn,
          short: '-w NUM_CONSUMERS',
          long: '--warn NUM_CONSUMERS',
@@ -124,7 +130,11 @@ class CheckRabbitMQConsumers < Sensu::Plugin::Check::CLI
 
   def run
     # create arrays to hold failures
-    missing = config[:queue] || []
+    missing = if config[:regex]
+                []
+              else
+                config[:queue] || []
+              end
     critical = []
     warn = []
 
@@ -132,7 +142,9 @@ class CheckRabbitMQConsumers < Sensu::Plugin::Check::CLI
       rabbit.queues.each do |queue|
         # if specific queues were passed only monitor those.
         # if specific queues to exclude were passed then skip those
-        if config[:queue]
+        if config[:regex]
+          next unless queue['name'] =~ /#{config[:queue].first}/
+        elsif config[:queue]
           next unless config[:queue].include?(queue['name'])
         elsif config[:exclude]
           next if config[:exclude].include?(queue['name'])