+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## [Unreleased]
 ### Added
+- check-rabbitmq-consumers.rb: Added ability to select queues with regular expressions. (@jtduby)
+
+### Added
 - ruby 2.4 support (@majormoses)
 
 ### Fixed